Distances

Distances

Overview

The following is a reference on all the distances defined in this package. Once a distance is defined, it can be evaluated in one of two ways, defined by the Distances.jl interface

dist = DistanceType(options)
d = evaluate(d, x1, x2)
d = dist(x1, x2)

Before we proceed, the following distances are available

SpectralDistances.AbstractDistance

SpectralDistances.AbstractRationalDistance

SpectralDistances.AbstractSignalDistance

SpectralDistances.BuresDistance

SpectralDistances.CoefficientDistance

SpectralDistances.DiscretizedRationalDistance

SpectralDistances.EnergyDistance

SpectralDistances.EuclideanRootDistance

SpectralDistances.HungarianRootDistance

SpectralDistances.KernelWassersteinRootDistance

SpectralDistances.ModelDistance

SpectralDistances.OptimalTransportHistogramDistance

SpectralDistances.RationalCramerDistance

SpectralDistances.RationalOptimalTransportDistance

SpectralDistances.SinkhornRootDistance

SpectralDistances.WelchOptimalTransportDistance

Some of these distances operate directly on signals, these are

EnergyDistance
ModelDistance
WelchOptimalTransportDistance

Of these, ModelDistance is a bit special, works like this

SpectralDistances.ModelDistance — Type.

ModelDistance{D <: AbstractDistance} <: AbstractSignalDistance

A model distance operates on signals and works by fitting an LTI model to the signals before calculating the distance. The distance between the LTI models is defined by the field distance. This is essentially a wrapper around the inner distance that handles the fitting of a model to the signals. How the model is fit is determined by fitmethod.

Arguments:

fitmethod::FitMethod: LS, TLS or PLR

distance::D: The inner distance between the models

Example:

using SpectralDistances
innerdistance = SinkhornRootDistance(domain=Continuous(), β=0.005, p=2)
dist = ModelDistance(TLS(na=30), innerdistance)

source

The inner distance in ModelDistance can be any AbstractRationalDistance. The options are

DiscretizedRationalDistance
RationalCramerDistance
RationalOptimalTransportDistance
SpectralDistances.AbstractCoefficientDistance
SpectralDistances.AbstractRootDistance

These distances operate on LTI models. Some operate on the coefficients of the models

CoefficientDistance

and some operate on the roots of the models

EuclideanRootDistance
HungarianRootDistance
KernelWassersteinRootDistance
SinkhornRootDistance

A full example

To use the SinkhornRootDistance and let it operate on signals, we may construct our distance object as follows

julia> innerdistance = SinkhornRootDistance(domain=Continuous(), β=0.005, p=2)
SinkhornRootDistance{Continuous,typeof(identity),Base.var"#56#57"{typeof(s1),typeof(residueweight)}}(Continuous(), identity, Base.var"#56#57"{typeof(s1),typeof(residueweight)}(SpectralDistances.s1, SpectralDistances.residueweight), 0.005, 10000, 2)

julia> dist = ModelDistance(TLS(na=30), innerdistance)
ModelDistance{SinkhornRootDistance{Continuous,typeof(identity),Base.var"#56#57"{typeof(s1),typeof(residueweight)}}}(TLS(30), SinkhornRootDistance{Continuous,typeof(identity),Base.var"#56#57"{typeof(s1),typeof(residueweight)}}(Continuous(), identity, Base.var"#56#57"{typeof(s1),typeof(residueweight)}(SpectralDistances.s1, SpectralDistances.residueweight), 0.005, 10000, 2))

julia> X1, X2 = randn(1000), randn(1000);

julia> dist(X1,X2)
┌ Warning: Roots on negative real axis, no corresponding continuous time representation exists.
└ @ SpectralDistances ~/build/baggepinnen/SpectralDistances.jl/src/ltimodels.jl:56
0.09224411089999843

julia> dist = ModelDistance(LS(na=2), innerdistance);

julia> t = 0:0.01:10;

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*1.1 .*t); # Two signals that are close in frequency

julia> dist(X1,X2)
0.0005614882602620935

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*2 .*t);   # Two signals that are further apart in frequency

julia> dist(X1,X2)
0.004004572174828294

Using Welch periodograms

We can calculate the Wasserstein distance between spectra estimated using the Welch method like so

julia> dist = WelchOptimalTransportDistance(p=2)
WelchOptimalTransportDistance{Nothing,Tuple{},NamedTuple{(),Tuple{}}}(nothing, 0.01, 10000, (), NamedTuple(), 2)

julia> X1, X2 = randn(1000), randn(1000);

julia> dist(X1,X2)
0.00013034949409317646

julia> t = 0:0.01:10;

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*1.1 .*t); # Two signals that are close in frequency

julia> dist(X1,X2)
8.823043208878249e-5

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*2 .*t);   # Two signals that are further apart in frequency

julia> dist(X1,X2)
0.00029130696022822513

Function reference

SpectralDistances.IPOT

SpectralDistances.discrete_grid_transportplan

SpectralDistances.domain

SpectralDistances.domain_transform

SpectralDistances.precompute

SpectralDistances.sinkhorn

Docstrings

SpectralDistances.AbstractDistance — Type.

The top level distance type

source

SpectralDistances.AbstractRationalDistance — Type.

All subtypes of this type operates on rational transfer functions

source

SpectralDistances.AbstractSignalDistance — Type.

All subtypes of this type operates on signals

source

SpectralDistances.BuresDistance — Type.

BuresDistance <: AbstractDistance

Distance between pos.def. matrices

source

SpectralDistances.CoefficientDistance — Type.

CoefficientDistance{D, ID} <: AbstractCoefficientDistance

Distance metric based on model coefficients

Arguments:

domain::D: Discrete or Continuous

distance::ID = SqEuclidean(): Inner distance between coeffs

source

SpectralDistances.DiscretizedRationalDistance — Type.

DiscretizedRationalDistance{WT, DT} <: AbstractRationalDistance

This distance discretizes the spectrum before performing the calculations.

Arguments:

w::WT = LinRange(0.01, 0.5, 300): Frequency set

distmat::DT = distmat_euclidean(w, w): DESCRIPTION

source

SpectralDistances.EnergyDistance — Type.

EnergyDistance <: AbstractSignalDistance

std(x1) - std(x2) This distance can be added to a loss function to ensure that the energy in the two signals is the same. Some of the optimal transport-based distances are invariant to the energy in the signal, requiring this extra cost if that invariance is not desired. Combining distances is done by putting two or more in a tuple. Usage: combined_loss = (primary_distance, EnergyDistance())

source

SpectralDistances.EuclideanRootDistance — Type.

EuclideanRootDistance{D, A, F1, F2} <: AbstractRootDistance

Simple euclidean distance between roots of transfer functions

Arguments:

domain::D: Discrete or Continuous

assignment::A = SortAssignement(imag): Determines how roots are assigned. An alternative is HungarianAssignement

transform::F1 = identity: DESCRIPTION

weight : A function used to calculate weights for the induvidual root distances. A good option is residueweight

p::Int = 2 : Order of the distance

source

SpectralDistances.HungarianRootDistance — Type.

HungarianRootDistance{D, ID <: Distances.PreMetric, F} <: AbstractRootDistance

Similar to EuclideanRootDistance but does the pole assignment using the Hungarian method.

Arguments:

domain::D: Discrete or Continuous

distance::ID = SqEuclidean(): Inner distance

transform::F = identity: If provided, this Function transforms all roots before the distance is calculated

source

SpectralDistances.KernelWassersteinRootDistance — Type.

KernelWassersteinRootDistance{D, F, DI} <: AbstractRootDistance

A kernel version of the root distance

Arguments:

domain::D: Discrete or Continuous

λ::Float64 = 1.0: Kernel precision, lower value means wider kernel.

transform::F = identity: If provided, this Function transforms all roots before the distance is calculated

distance::DI = SqEuclidean(): Inner distance

source

SpectralDistances.OptimalTransportHistogramDistance — Type.

OptimalTransportHistogramDistance{DT} <: AbstractDistance

What it sounds like

Arguments:

p::Int = 1: order

source

SpectralDistances.RationalCramerDistance — Type.

RationalCramerDistance{DT} <: AbstractRationalDistance

Similar to RationalOptimalTransportDistance but does not use inverse functions.

Arguments:

domain::DT: Discrete or Continuous

p::Int = 2: order

interval = (-(float(π)), float(π)): Integration interval

source

SpectralDistances.RationalOptimalTransportDistance — Type.

RationalOptimalTransportDistance{DT, MT} <: AbstractRationalDistance

calculates the Wasserstein distance using the closed-form solution based on integrals and inverse cumulative functions.

Arguments:

domain::DT: Discrete or Continuous

p::Int = 1: order

magnitude::MT = Identity():

interval = (-(float(π)), float(π)): Integration interval

source

SpectralDistances.SinkhornRootDistance — Type.

SinkhornRootDistance{D, F1, F2} <: AbstractRootDistance

The Sinkhorn distance between roots. The weights are provided by weight, which defaults to residueweight.

Arguments:

domain::D: Discrete or Continuous

transform::F1 = identity: Probably not needed.

weight::F2 =s1 ∘ residueweight: A function used to calculate weights for the induvidual root distances.

β::Float64 = 0.01: Amount of entropy regularization

iters::Int = 10000: Number of iterations of the Sinkhorn algorithm.

p::Int = 2 : Order of the distance

source

SpectralDistances.WelchOptimalTransportDistance — Type.

WelchOptimalTransportDistance{DT, AT <: Tuple, KWT <: NamedTuple} <: AbstractSignalDistance

Calculates the Wasserstein distance between two signals by estimating a Welch periodogram of each.

Arguments:

distmat::DT: you may provide a matrix array for this

args::AT = (): Options to the Welch function

kwargs::KWT = NamedTuple(): Options to the Welch function

p::Int = 2 : Order of the distance

source

SpectralDistances.discrete_grid_transportplan — Method.

discrete_grid_transportplan(x::AbstractVector{T}, y::AbstractVector{T}, tol=sqrt(eps(T))) where T

Calculate the optimal-transport plan between two vectors that are assumed to have the same support, with sorted support points.

source

SpectralDistances.domain — Method.

domain(d::AbstractDistance)

Return the domain of the distance

source

SpectralDistances.domain_transform — Method.

domain_transform(d::AbstractDistance, e)

Change domain of roots

source

SpectralDistances.precompute — Function.

precompute(d::AbstractDistance, As, threads=true)

Perform computations that only need to be donce once when several pairwise distances are to be computed

Arguments:

As: A vector of models

threads: Us multithreading? (true)

source

SpectralDistances.IPOT — Method.

γ, u, v = IPOT(C, a, b; β=1, iters=1000)

The Inexact Proximal point method for exact Optimal Transport problem (IPOT) (Sinkhorn-like) algorithm. C is the cost matrix and a,b are vectors that sum to one. Returns the optimal plan and the dual potentials. See also sinkhorn. β does not have to go to 0 for this alg to return the optimal distance.

A Fast Proximal Point Method for Computing Exact Wasserstein Distance Yujia Xie, Xiangfeng Wang, Ruijia Wang, Hongyuan Zha https://arxiv.org/abs/1802.04307

source

SpectralDistances.sinkhorn — Method.

γ, u, v = sinkhorn(C, a, b; β=1e-1, iters=1000)

The Sinkhorn algorithm. C is the cost matrix and a,b are vectors that sum to one. Returns the optimal plan and the dual potentials. See also IPOT.

source