Interpolations and Barycenters

Some distances distance define the existence of a shortest path, a geodesic. An interpolation is essentially a datapoint on that shortest path. We provide some functionality to interpolate between different spectra and models under transport-based metrics.

Below is an example usage of interpolations. We initially create two random systems, we then define the distance under which to interpolate and then calculate the frequency response for some different values of the interpolation parameter $t \in (0,1)$

using SpectralDistances, ControlSystems, Distances, Plots, Random
plotly()
Random.seed!(0)

n = 4
r1 = complex.(-0.01 .+ 0.001randn(3), 2randn(3))
r1 = ContinuousRoots([r1; conj.(r1)])

r2 = complex.(-0.01 .+ 0.001randn(3), 2randn(3))
r2 = ContinuousRoots([r2; conj.(r2)])

r1,r2 = normalize_energy.((r1, r2))

A1 = AR(r1)
A2 = AR(r2)

##
fig1   = plot()
t      = 0.1
dist   = RationalOptimalTransportDistance(domain=Continuous(), p=2, interval=(0., exp10(1.01)))
interp = SpectralDistances.interpolator(dist, A1, A2)
w      = exp10.(LinRange(-1.5, 1, 300))
for t = LinRange(0, 1, 7)
    Φ = clamp.(interp(w,t), 1e-10, 100)
    plot!(w, sqrt.(Φ), xscale=:log10, yscale=:log10, line_z = t, lab="", xlabel="", title="W_2", ylims=(1e-3, 1e1), colorbar=false, l=(1,), c=:viridis)
end

rdist  = EuclideanRootDistance(domain                           =Continuous(), p=2)
interp = SpectralDistances.interpolator(rdist, A1, A2, normalize=false)
fig2   = plot()
for t = LinRange(0, 1, 7)
    Φ = interp(w,t)
    plot!(w, sqrt.(Φ), xscale=:log10, yscale=:log10, line_z = t, lab="", xlabel="", title="RD", ylims=(1e-3, 1e1), colorbar=false, l=(1,), c=:viridis)
end

fig3 = plot()
Φ1   = bode(tf(A1), w)[1][:]
Φ2   = bode(tf(A2), w)[1][:]
for t = LinRange(0, 1, 7)
    plot!(w, (1-t).*Φ1 .+ t.*Φ2, xscale=:log10, yscale=:log10, line_z = t, lab="", xlabel="Frequency", title="L_2", ylims=(1e-3, 1e1), colorbar=false, l=(1,), c=:viridis)
end

fig = plot(fig1, fig2, fig3, layout=(3,1))

Barycenters

A barycenter is a generalization the the arithmetic mean to metrics other than the Euclidean. A barycenter between models is calculated like this

bc = barycenter(distance, models)

It can be useful to provide some options to the solvers:

options = (solver=sinkhorn_log!, tol=1e-8, iters=1_000_000, γ=0.0, uniform=true, inneriters=500_000, innertol=1e-6)
distance = OptimalTransportRootDistance(domain=Continuous(), p=2, β=0.01, weight=simplex_residueweight)
bc = barycenter(distance, models; options...)

We can plot the barycenters:

using SpectralDistances, ControlSystems, Plots
models   = examplemodels(3)
distance = OptimalTransportRootDistance(domain=Continuous())
bc       = barycenter(distance, models)
w        = exp10.(LinRange(-0.5, 0.5, 350)) # Frequency vector
G        = tf.(models) # Convert models to transfer functions from ControlSystems.jl
plot()
bodeplot!.(G, Ref(w), plotphase=false, lab="Input models", linestyle=:auto)
bodeplot!(tf(bc), w, plotphase=false, lab="Barycenter", xscale=:identity, c=:green)

K-Barycenters

Below, we show an example of how one can run the K-barycenter algorithm on a collection of sound signals. sounds is expected to be of type Vector{Vector{T}}. The example further assumes that there is a vector of labels::Vector{Int} that contain the true classes of the datapoints, which you do not have in an unsupervised setting.

using SpectralDistances, ControlSystems
fitmethod = TLS(na=12)
models = SpectralDistances.fitmodel.(fitmethod, sounds)
G = tf.(models) # Convert to transfer functions for visualization etc.

##
using Clustering
dist = OptimalTransportRootDistance(domain=Continuous(), β=0.01, weight=simplex_residueweight)
@time clusterresult = SpectralDistances.kbarycenters(
    dist,
    models,
    n_classes, # number of clusters
    seed    = :rand,
    solver  = sinkhorn_log!,
    tol     = 2e-6,
    innertol = 2e-6,
    iters   = 100000,
    inneriters = 100000,
    verbose = true,
    output  = :best,
    uniform = true,
    kiters  = 10
)

bc,ass = clusterresult.barycenters, clusterresult.assignments

# Visualize results
using MLBase, Plots.PlotMeasures, AudioClustering
newass,perm = AudioClustering.associate_clusters(labels,ass)
classinds   = 1:n_classes
yt          = (classinds, [label_strings[findfirst(labels .== i)] for i in classinds])

@show mean(labels .== newass)
cm = confusmat(n_classes,labels,newass)
heatmap(cm./sum(cm,dims=2), xlabel="Cluster assignment",ylabel="Best matching class", color=:viridis)
anns = [(reverse(ci.I)..., text(val,12,:gray)) for (ci,val) in zip(CartesianIndices(cm)[:], vec(cm))]
annotate!(anns)
yticks!(yt)
xticks!(yt, xrotation=45)
current()

The figure should look like the last figure in the paper.

SpectralDistances.ISA

SpectralDistances.barycenter

SpectralDistances.barycenter

SpectralDistances.barycenter

SpectralDistances.barycentric_coordinates

SpectralDistances.embedding

SpectralDistances.interpolator

SpectralDistances.kbarycenters

SpectralDistances.kbarycenters

SpectralDistances.interpolator — Method

interpolator(d, A1, A2)

Perform displacement interpolation between two models.

SpectralDistances.ISA — Function

ISA(X, w = nothing; iters = 100, printerval = typemax(Int))

Iterative swapping algorithm from "On the Computation of Wasserstein barycenters", Giovanni Puccetti et al.

Arguments:

X: vector of d×k matrices where d is dimension and k number of atoms

w: weights. See the files test_barycenter.jl for different uses.

iters: maximum number of iterations

printerval: print this often

SpectralDistances.barycenter — Function

barycenter(d::OptimalTransportRootDistance, models; normalize = true, kwargs...)

barycenter(d, models)
barycenter(d, models, λ; normalize, uniform, solver, kwargs...)

Approximately calculate the barycenter supported on the same number of atoms as the number of poles in the models.

The solver can be selected by providing a keword argument, example: solver=IPOT.

Uses the algorithms from "Fast Computation of Wasserstein Barycenters"

Example:

models = examplemodels(10)

d = OptimalTransportRootDistance(domain=SpectralDistances.Continuous(),p=2, weight=residueweight, β=0.01)
Xe = barycenter(d, models, solver=sinkhorn_log!)

plot()
pzmap!.(models)
pzmap!(tf(Xe), m=:c, title="Barycenter OptimalTransportRootDistance", lab="BC")
current()

Arguments:

models: vector of AR models

normalize: make sure weights sum to 1

kwargs: are sent to the solver

SpectralDistances.barycenter — Method

barycenter(X::Vector{<:AbstractArray}, λ)

Calculate the weighted barycenter for point clouds in X. Each X[i] has the shame n_dims × n_atoms λ is the weight vector that should sum to 1.

SpectralDistances.barycenter — Method

barycenter(d::EuclideanRootDistance, models::AbstractVector)

Example:

models = examplemodels(10)

Xe = barycenter(EuclideanRootDistance(domain=SpectralDistances.Continuous(),p=2), models)

G = tf.(models)
plot()
pzmap!.(G)
pzmap!(tf(Xe), m=:c, title="Barycenter EuclideanRootDistance")
current()

SpectralDistances.barycentric_coordinates — Method

λ = barycentric_coordinates(pl, ql, p, q; options, kwargs...)

Compute the barycentric coordinates λ such that sum(λᵢ W(pᵢ,q) for i in eachindex(p)) is minimized.

This function works best with the sinkhorn_log! solver, a large β (around 1) and small tolerance. These are set using kwargs....

Arguments:

pl: Atoms in measures p, vector, length n_measures, of matrices of size n_dims × n_atoms

ql: Atoms in measure q

p: Measures p, a matrix of weight vectors, size n_atoms × n_measures that sums to 1

q: the veight vector for measure q, length is n_atoms

options: For the Optim solver. Defaults are options = Optim.Options(store_trace=false, show_trace=false, show_every=0, iterations=20, allow_f_increases=true, time_limit=100, x_tol=1e-5, f_tol=1e-6, g_tol=1e-6, f_calls_limit=0, g_calls_limit=0)

solver: = sinkhorn_log! solver

tol: = 1e-7 tolerance

β: = 0.1 entropy regularization. This function works best with rather large regularization, hence the large default value.

kwargs: these are sent to the solver algorithm.

Example:

using SpectralDistances, ControlSystems, Optim
models = examplemodels(10)

d = OptimalTransportRootDistance(
    domain = SpectralDistances.Continuous(),
    p      = 2,
    weight = residueweight,
    β      = 0.01,
)
Xe = barycenter(d, models, solver=sinkhorn_log!)

G = tf.(models)
plot()
pzmap!.(G)
pzmap!(tf(Xe), m=:c, title="Barycenter OptimalTransportRootDistance", lab="BC")

options = Optim.Options(store_trace       = true,
                        show_trace        = false,
                        show_every        = 1,
                        iterations        = 50,
                        allow_f_increases = true,
                        time_limit        = 100,
                        x_tol             = 1e-7,
                        f_tol             = 1e-7,
                        g_tol             = 1e-7,
                        f_calls_limit     = 0,
                        g_calls_limit     = 0)

method = LBFGS()
λ = barycentric_coordinates(d, models, Xe, method,
    options = options,
    solver  = sinkhorn_log!,
    robust  = true,
    uniform = true,
    tol     = 1e-6,
)
bar(λ, title="Barycentric coorinates")

G = tf.(models)
plot()
pzmap!.(G, lab="")
pzmap!(tf(Xe), m = :c, title = "Barycenter OptimalTransportRootDistance", lab = "BC")
# It's okay if the last system dot does not match the barycenter exactly, there are limited models to choose from.
pzmap!(G[argmax(λ)], m = :c, lab = "Largest bc coord", legend = true)

SpectralDistances.embedding — Method

embedding([::Type{Vector}], m, [full=true])

Returns a Vector/Matrix containing the roots of m. full indicates whether or not to use all poles or only one half-plane.

SpectralDistances.kbarycenters — Method

kbarycenters(X, p, k; seed = :rand, kiters = 10, verbose = false, output = :best, kwargs...)

Clustering using K-barycenters. If you want to cluster spectra, consider the method that accepts models instead.

Arguments:

X: Support of input measures, Vector{Matrix} where each matrix is ndims×natoms

p: Weights of input measures Vector{Vector} where each matrix is of length n_atoms and should sum to 1.

k: number of clusters

seed: :rand or :eq

kiters: number of iterations

verbose: print stuff?

output: output lowest cost clustering or :latest?

kwargs: are sent to the inner solvers.

SpectralDistances.kbarycenters — Method

kbarycenters(d::OptimalTransportRootDistance, models::Vector{<:AbstractModel}, k; normalize = true, kwargs...)

This function is only available if using Clustering.

Example:

clusterresult = kbarycenters(
    dist,
    models,
    k,         # number of clusters
    seed       = :rand,
    solver     = sinkhorn_log!,
    tol        = 2e-6,
    innertol   = 2e-6,
    iters      = 100000,
    inneriters = 100000,
    verbose    = true,
    output     = :best,
    uniform    = true,
    kiters     = 10
)

The docs contain a more detailed example

Arguments:

models: A vector of models

k: number of clusters

normalize: Whether or not to normalize the weight vectors (recommended)

kwargs: are sent to inner solvers, (solver,tol,iters)