Plotting

models = examplemodels(2)
assignmentplot(models[1], models[2], d=Discrete(), p=2)

dist = OptimalTransportRootDistance(domain=Continuous())
flowplot(dist, models...)

Apart from the functions above, the plotting facilities from ControlSystems.jl should also work on models from this package, e.g., bodeplot, pzmap etc.

SpectralDistances.assignmentplot

SpectralDistances.flowplot

SpectralDistances.assignmentplot — Function

assignmentplot(m1, m2)

Plots the poles of two models and the optimal assignment between them

SpectralDistances.flowplot — Function

flowplot(dist,m1, m2)

Plots the poles of two models and the optimal mass flow between them