Distances

Overview

The following is a reference for all the distances defined in this package. Once a distance object is defined, it can be evaluated in one of two ways, defined by the Distances.jl interface

dist = DistanceType(options)
d = evaluate(d, x1, x2; kwargs...) # keyword arguments are used to control the solvers for some transport-based distances
d = dist(x1, x2) # A shorter syntax for calling the distance

Note: All distances return the distance raised to the power p, thus RationalOptimalTransportDistance(p=2)(x1,x2) == W₂(x1,x2)^2 where W₂ denotes the Wasserstein distance of order 2.

Before we proceed, we list a number of classes of distances that are available

SpectralDistances.AbstractDistance

SpectralDistances.AbstractRationalDistance

SpectralDistances.AbstractSignalDistance

SpectralDistances.BuresDistance

SpectralDistances.CoefficientDistance

SpectralDistances.DiscretizedRationalDistance

SpectralDistances.EnergyDistance

SpectralDistances.EuclideanRootDistance

SpectralDistances.HungarianRootDistance

SpectralDistances.KernelWassersteinRootDistance

SpectralDistances.ModelDistance

SpectralDistances.OptimalTransportHistogramDistance

SpectralDistances.OptimalTransportRootDistance

SpectralDistances.RationalCramerDistance

SpectralDistances.RationalOptimalTransportDistance

SpectralDistances.WelchLPDistance

SpectralDistances.WelchOptimalTransportDistance

Some of these distances operate directly on signals, these are

EnergyDistance
ModelDistance
SpectralDistances.AbstractWelchDistance

Of these, ModelDistance is a bit special, works like this

SpectralDistances.ModelDistance — Type

ModelDistance{D <: AbstractDistance} <: AbstractSignalDistance

A model distance operates on signals and works by fitting an LTI model to the signals before calculating the distance. The distance between the LTI models is defined by the field distance. This is essentially a wrapper around the inner distance that handles the fitting of a model to the signals. How the model is fit is determined by fitmethod.

Arguments:

fitmethod::FitMethod: LS, TLS or PLR

distance::D: The inner distance between the models

Example:

using SpectralDistances
innerdistance = OptimalTransportRootDistance(domain=Continuous(), β=0.005, p=2)
dist = ModelDistance(TLS(na=30), innerdistance)

The inner distance in ModelDistance can be any AbstractRationalDistance. The options are

DiscretizedRationalDistance
RationalCramerDistance
RationalOptimalTransportDistance
SpectralDistances.AbstractCoefficientDistance
SpectralDistances.AbstractRootDistance

These distances operate on LTI models. Some operate on the coefficients of the models

CoefficientDistance

and some operate on the roots of the models

EuclideanRootDistance
HungarianRootDistance
KernelWassersteinRootDistance
OptimalTransportRootDistance

A full example

To use the OptimalTransportRootDistance and let it operate on signals, we may construct our distance object as follows

julia> innerdistance = OptimalTransportRootDistance(domain=Continuous(), β=0.005, p=2)
OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Nothing}(Continuous(), identity, SpectralDistances.simplex_residueweight, 0.005, 2, nothing)

julia> dist = ModelDistance(TLS(na=10), innerdistance)
ModelDistance{OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Nothing},TLS}(TLS(10, 0.0), OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Nothing}(Continuous(), identity, SpectralDistances.simplex_residueweight, 0.005, 2, nothing))

julia> X1, X2 = randn(1000), randn(1000);

julia> dist(X1,X2)
0.020900305881628594

julia> dist = ModelDistance(LS(na=2), innerdistance);

julia> t = 0:0.01:10;

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*1.1 .*t); # Two signals that are close in frequency

julia> dist(X1,X2)
0.0005678342162688203

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*2 .*t);   # Two signals that are further apart in frequency

julia> dist(X1,X2)
0.00407322855228585

Using Welch periodograms

We can calculate the Wasserstein distance between spectra estimated using the Welch method like so

julia> dist = WelchOptimalTransportDistance(p=2)
WelchOptimalTransportDistance{Nothing,Tuple{},NamedTuple{(),Tuple{}}}(nothing, (), NamedTuple(), 2)

julia> X1, X2 = randn(1000), randn(1000);

julia> dist(X1,X2)
0.0003535845316060714

julia> t = 0:0.01:10;

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*1.1 .*t); # Two signals that are close in frequency

julia> dist(X1,X2)
8.823043208878249e-5

julia> X1, X2 = sin.(2π*1 .*t), sin.(2π*2 .*t);   # Two signals that are further apart in frequency

julia> dist(X1,X2)
0.00029130696022822513

Gradients

Some distances will allow you to propagate gradients through them. Below is an example using Zygote and the OptimalTransportRootDistance

using Zygote
Zygote.@nograd rand # Currently needed woraround
x1 = SpectralDistances.bp_filter(randn(100000), (0.1,0.3))  # Create two signals
x2 = SpectralDistances.bp_filter(randn(100000), (0.1,0.2))
fm = LS(na=10)      # LS is the best supported fitmethod for gradients

dist = ModelDistance(fm,OptimalTransportRootDistance(domain = Continuous()))      # Since we're measureing distance between signals, we wrap the distance in a ModelDistance

ModelDistance{OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Nothing},LS}(LS(10, 0.01), OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Nothing}(Continuous(), identity, SpectralDistances.simplex_residueweight, 0.01, 2, nothing))

julia> dist(x1,x2)
0.027235323880123966

julia> ∇x1 = Zygote.gradient(x->real(evaluate(dist,x,x2)), x1)[1] # The call to real is a workaround for a Zygote bug
100000-element Array{Float64,1}:
  3.7941886583343173e-6
 -1.1432199189232572e-5
  1.1478231035920261e-5
 -1.5136927772373006e-5
  1.1005764826205785e-5
 -1.9363520276059635e-5
  1.1180874806529159e-5
 -2.222183588789598e-5
  3.974322670971006e-6
 -2.5128611110977215e-6
  ⋮
  2.1251308019700997e-6
 -7.620799881547004e-6
 -1.2178382792914463e-5
  1.076396830130313e-5
 -1.2636532102811444e-6
  1.6310837403546995e-5
 -1.594438043138127e-5
  6.951128303328147e-6
 -1.490294399838832e-5

The differentiation takes some time, but it should be fast enough to be generally useful for gradient-based learning of autoencoders etc. The following is a benchmark performed on an old laptop without GPU (the distances are not yet tested on GPUs)

@btime Zygote.gradient(x->real(evaluate($dist,x,$x2)), $x1);
#  134.965 ms (107566 allocations: 134.77 MiB)

with a precomputed reference model, it goes even faster

m2 = fm(x2)
m2 = change_precision(Float64, m2) # Tihs step is important for performance
@btime Zygote.gradient(x->real(evaluate($dist,x,$m2)), $x1);
#  80.200 ms (103437 allocations: 69.62 MiB)

The same benchmarks performed on a 2019 desktop computer yields the following timings

julia> @btime Zygote.gradient(x->real(evaluate($dist,x,$x2)), $x1);
  45.926 ms (107748 allocations: 136.18 MiB)

julia> @btime Zygote.gradient(x->real(evaluate($dist,x,$m2)), $x1);
  25.120 ms (103619 allocations: 71.03 MiB)

Unbalanced transport

There are situations in which one would like to avoid fully transporting all mass between two measures. A few such cases are

The two measures do not have the same mass. In this case, the standard, balanced, optimal-transport problem is unfeasible.

Energy is somehow lost or added to one spectra in a way that should not be accounted for by transport. This would be the case if

Spectral energy is absorbed by a channel through which a signal is propagated. In this case it would not make sense to try to transport mass from the other spectrum away from the absorbed (dampended) frequency.

Spectral energy is added by a noise source. This energy should ideally not be considered for transport and should rather be destroyed.

For situations like this, an AbstractDivergence can be supplied to the OptimalTransportRootDistance. This divergence specifies how expensive it is to create or destroy mass in the spectra. The available divergences are listed in the docs of UnbalancedOptimalTransport.jl, to which we outsource the solving of the unbalanced problem. For convenience, the wrapper sinkhorn_unbalanced is available to interface the unbalanced solver in the same way as the solvers from this package are interfaced.

julia> using DSP

julia> fm = LS(na = 10);

julia> m1 = fm(filtfilt(ones(10), [10], randn(1000)));
┌ Warning: Roots on negative real axis, no corresponding continuous time representation exists. Consider prefiltering the signal or decreasing the regularization factor.
└ @ SpectralDistances ~/work/SpectralDistances.jl/SpectralDistances.jl/src/ltimodels.jl:68

julia> m2 = fm(filtfilt(ones(5), [5], randn(1000)));
┌ Warning: Roots on negative real axis, no corresponding continuous time representation exists. Consider prefiltering the signal or decreasing the regularization factor.
└ @ SpectralDistances ~/work/SpectralDistances.jl/SpectralDistances.jl/src/ltimodels.jl:68

julia> dist = OptimalTransportRootDistance(domain = Continuous(), p=1, divergence=Balanced())
OptimalTransportRootDistance{Continuous,typeof(identity),typeof(simplex_residueweight),Balanced}(Continuous(), identity, SpectralDistances.simplex_residueweight, 0.01, 1, Balanced())

julia> d1 = evaluate(dist,m1,m2)
0.13548131784016387

julia> dist = OptimalTransportRootDistance(domain = Continuous(), p=1, divergence=KL(1.0));

julia> d2 = evaluate(dist,m1,m2)
0.12916171188372763

julia> dist = OptimalTransportRootDistance(domain = Continuous(), p=1, divergence=KL(10.0));

julia> d3 = evaluate(dist,m1,m2)
0.13481793967995637

julia> dist = OptimalTransportRootDistance(domain = Continuous(), p=1, divergence=KL(0.01));

julia> d4 = evaluate(dist,m1,m2)
0.029523569846449584

julia> d1 > d3 > d2 > d4
true

When the distance is evaluated the second time, unbalanced transport is used. The d2 should be equal to or smaller than d1. If we make the KL term larger, the distance approaches the balanced cost, and if we make it smaller, it becomes very cheap to create/destroy mass and less mass is transported.

The first case, where divergence=Balanced() was supplied, should be equivalent to not providing any divergence at all. In pratice results might differ slightly since a different solver implementation is used.

Currently, only distance calculations using unbalanced transport is supported. Barycenter calculations will currently ignore the divergence.

Below is an example in which the unbalanced transport between two systems is computed. The two systems do not have the same number of poles, and if destruction of mass is made cheap, not all mass is transported. The thickness of the lines indicate mass flow.

using Plots
m1 = AR(Continuous(), [1, 0.1, 1.3])                        |> change_precision(Float64)
m2 = AR(Continuous(), polyconv([1, 0.1, 1], [1, 0.1, 1.2])) |> change_precision(Float64)
D  = SpectralDistances.distmat_euclidean(m1.pc, m2.pc)
w1, w2 = unitweight.((m1, m2))
figs = map([0.001, 0.01, 0.1]) do tv
    divergence = TV(tv)
    Γ, a, b = sinkhorn_unbalanced(D, w1, w2, divergence, β = 0.01)
    lineS   = 20Γ
    lineS[lineS.<0.1] .= 0
    alphaS  = lineS ./ maximum(lineS)
    f       = scatter(m1.pc, legend = false, ms = 10, title = "TV=$tv")
    scatter!(m2.pc, ms = 10)
    for (i, p1) in enumerate(m1.pc), (j, p2) in enumerate(m2.pc)
        coords = [p1, p2]
        plot!(
            real(coords),
            imag(coords),
            linewidth = lineS[i, j],
            alpha     = alphaS[i, j],
            color     = :black,
        )
    end
    f
end
plot(figs..., layout = (1, 3), ylims = (0.9, 1.2))

┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283
┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283
┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283
┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283
┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283
┌ Warning: No strict ticks found
└ @ PlotUtils ~/.julia/packages/PlotUtils/10ojB/src/ticks.jl:283

Function reference

SpectralDistances.IPOT

SpectralDistances.discrete_grid_transportplan

SpectralDistances.domain

SpectralDistances.domain_transform

SpectralDistances.ot_jump

SpectralDistances.precompute

SpectralDistances.sinkhorn

SpectralDistances.sinkhorn_log

SpectralDistances.sinkhorn_log!

SpectralDistances.sinkhorn_unbalanced

Docstrings

SpectralDistances.AbstractDistance — Type

The top level distance type

SpectralDistances.AbstractRationalDistance — Type

All subtypes of this type operates on rational transfer functions

SpectralDistances.AbstractSignalDistance — Type

All subtypes of this type operates on signals

SpectralDistances.BuresDistance — Type

BuresDistance <: AbstractDistance

Distance between pos.def. matrices

SpectralDistances.CoefficientDistance — Type

CoefficientDistance{D, ID} <: AbstractCoefficientDistance

Distance metric based on model coefficients

Arguments:

domain::D: Discrete or Continuous

distance::ID = SqEuclidean(): Inner distance between coeffs

SpectralDistances.DiscretizedRationalDistance — Type

DiscretizedRationalDistance{WT, DT} <: AbstractRationalDistance

This distance discretizes the spectrum before performing the calculations.

Arguments:

w::WT = LinRange(0.01, 0.5, 300): Frequency set

distmat::DT = distmat_euclidean(w, w): DESCRIPTION

SpectralDistances.EnergyDistance — Type

EnergyDistance <: AbstractSignalDistance

std(x1) - std(x2) This distance can be added to a loss function to ensure that the energy in the two signals is the same. Some of the optimal transport-based distances are invariant to the energy in the signal, requiring this extra cost if that invariance is not desired. Combining distances is done by putting two or more in a tuple. Usage: combined_loss = (primary_distance, EnergyDistance())

SpectralDistances.EuclideanRootDistance — Type

EuclideanRootDistance{D, A, F1, F2} <: AbstractRootDistance

Simple euclidean distance between roots of transfer functions

Arguments:

domain::D: Discrete or Continuous

assignment::A = SortAssignement(imag): Determines how roots are assigned. An alternative is HungarianAssignement

transform::F1 = identity: DESCRIPTION

weight : A function used to calculate weights for the induvidual root distances. A good option is residueweight

p::Int = 2 : Order of the distance

SpectralDistances.HungarianRootDistance — Type

HungarianRootDistance{D, ID <: Distances.PreMetric, F} <: AbstractRootDistance

Similar to EuclideanRootDistance but does the pole assignment using the Hungarian method.

Arguments:

domain::D: Discrete or Continuous

distance::ID = SqEuclidean(): Inner distance

transform::F = identity: If provided, this Function transforms all roots before the distance is calculated

SpectralDistances.KernelWassersteinRootDistance — Type

KernelWassersteinRootDistance{D, F, DI} <: AbstractRootDistance

A kernel version of the root distance

Arguments:

domain::D: Discrete or Continuous

λ::Float64 = 1.0: Kernel precision, lower value means wider kernel.

transform::F = identity: If provided, this Function transforms all roots before the distance is calculated

distance::DI = SqEuclidean(): Inner distance

SpectralDistances.OptimalTransportHistogramDistance — Type

OptimalTransportHistogramDistance{DT} <: AbstractDistance

Optimal transport between two histograms. If you pass two vectors to this distance two histograms will be computed automatically (fit(Histogram,x)). Pass two histograms to get better manual control.

Arguments:

p::Int = 1: order

SpectralDistances.OptimalTransportRootDistance — Type

OptimalTransportRootDistance{D, F1, F2, S} <: AbstractRootDistance

The Sinkhorn distance between roots. The weights are provided by weight, which defaults to residueweight.

Arguments:

domain::D: Discrete or Continuous

transform::F1 = identity: Probably not needed.

weight::F2 =simplex_residueweight: A function used to calculate weights for the induvidual root distances.

β::Float64 = 0.01: Amount of entropy regularization

p::Int = 2 : Order of the distance

divergence::S = nothing: A divergence that penalizes creation and destruction of mass.

SpectralDistances.RationalCramerDistance — Type

RationalCramerDistance{DT} <: AbstractRationalDistance

Similar to RationalOptimalTransportDistance but does not use inverse functions.

Arguments:

domain::DT: Discrete or Continuous

p::Int = 2: order

interval = (-(float(π)), float(π)): Integration interval

SpectralDistances.RationalOptimalTransportDistance — Type

RationalOptimalTransportDistance{DT, MT} <: AbstractRationalDistance

calculates the Wasserstein distance using the closed-form solution based on integrals and inverse cumulative functions.

Arguments:

domain::DT: Discrete or Continuous

p::Int = 1: order

magnitude::MT = Identity():

interval = (-(float(π)), float(π)): Integration interval

SpectralDistances.WelchLPDistance — Type

WelchLPDistance{AT <: Tuple, KWT <: NamedTuple, F} <: AbstractWelchDistance

Lᵖ distance between welch spectra, mean(abs(x1-x2)^p).

#Arguments:

args::AT = (): These are sent to welch_pgram

kwargs::KWT = NamedTuple(): These are sent to welch_pgram

p::Int = 2: Order of the distance

normalized::Bool = true: Normlize spectrum to sum to 1 (recommended)

transform::F = identity: Optional function to apply to the spectrum, example log1p or sqrt. Must not produce negative values, so log is not a good idea. The function is applied like this: transform.(x1).

SpectralDistances.WelchOptimalTransportDistance — Type

WelchOptimalTransportDistance{DT, AT <: Tuple, KWT <: NamedTuple} <: AbstractWelchDistance

Calculates the Wasserstein distance between two signals by estimating a Welch periodogram of each.

Arguments:

distmat::DT: you may provide a matrix array for this

args::AT = (): Options to the Welch function

kwargs::KWT = NamedTuple(): Options to the Welch function

p::Int = 2 : Order of the distance

SpectralDistances.discrete_grid_transportplan — Method

discrete_grid_transportplan(x::AbstractVector{T}, y::AbstractVector{T}, tol=sqrt(eps(T))) where T

Calculate the optimal-transport plan between two vectors that are assumed to have the same support, with sorted support points.

SpectralDistances.domain — Method

domain(d::AbstractDistance)

domain(d::Any) -> Discrete

Return the domain of the distance

SpectralDistances.domain_transform — Method

domain_transform(d::AbstractDistance, e)

Change domain of roots

SpectralDistances.precompute — Function

precompute(d::AbstractDistance, As, threads=true)

Perform computations that only need to be donce once when several pairwise distances are to be computed

Arguments:

As: A vector of models

threads: Us multithreading? (true)

SpectralDistances.IPOT — Method

Γ, u, v = IPOT(C, a, b; β=1, iters=1000)

The Inexact Proximal point method for exact Optimal Transport problem (IPOT) (Sinkhorn-like) algorithm. C is the cost matrix and a,b are vectors that sum to one. Returns the optimal plan and the dual potentials. See also sinkhorn. β does not have to go to 0 for this alg to return the optimal distance, in fact, if β is set too low, this alg will encounter numerical problems.

A Fast Proximal Point Method for Computing Exact Wasserstein Distance Yujia Xie, Xiangfeng Wang, Ruijia Wang, Hongyuan Zha https://arxiv.org/abs/1802.04307

SpectralDistances.sinkhorn — Method

Γ, u, v = sinkhorn(C, a, b; β=1e-1, iters=1000)

The Sinkhorn algorithm. C is the cost matrix and a,b are vectors that sum to one. Returns the optimal plan and the dual potentials. This function is relatively slow, see also sinkhorn_log! IPOT and sinkhorn_log for faster algorithms.

SpectralDistances.sinkhorn_log! — Method

Same as sinkhorn_log but operates in-place to save memory allocations. This function has higher performance than sinkhorn_log, but might not work as well with AD libraries.

SpectralDistances.sinkhorn_log — Method

Γ, u, v = sinkhorn_log(C, a, b; β=1e-1, iters=1000, tol=1e-8)

The Sinkhorn algorithm (log-stabilized). C is the cost matrix and a,b are vectors that sum to one. Returns the optimal plan and the dual potentials. See also sinkhorn_log! for a faster implementation operating in-place, and IPOT for a potentially more exact solution.

When this function is being differentiated, warnings about inaccurate solutions are turned off. You may choose to manually asses the error in the constrains by ea, eb = SpectralDistances.ot_error(Γ, a, b).

The IPOT algorithm: https://arxiv.org/pdf/1610.06519.pdf

SpectralDistances.sinkhorn_unbalanced — Method

Γ, u, v = sinkhorn_unbalanced(C, a, b, divergence; β=1e-1, iters=1000, tol=1e-8)

The Unbalanced Sinkhorn algorithm (log-stabilized). C is the cost matrix and a,b are vectors that are not required to sum to one.

Ref: "Sinkhorn Divergences for Unbalanced Optimal Transport" https://arxiv.org/abs/1910.12958 Makes use of UnbalancedOptimalTransport.jl

SpectralDistances.ot_jump — Method

ot_jump(D, P1, P2)

Solve the optimal transport problem using JuMP. This function is only available if using JuMP, GLPK.

Arguments:

D: Distance matrix

P1: Weight vector 1

P2: Weight vector 2

Details

Transport-based distances may require some tuning parameters to be set for the solvers. The available solvers are

sinkhorn: not recommended due to numerical issues, but this is the most commonly cited algorithm.

sinkhorn_log: better numerical stability than the standard.

sinkhorn_log!: in-place version that is faster, but some AD libraries might not like it (often the default if no solver is provided).

sinkhorn_unbalanced: this solver accepts a divergence that penalizes creation/destruction of mass. It thus handles measure of different masses and can choose to create/destroy mass instead of transporting it.

IPOT Finds exact solution (without entropy regularization), requires β around 0.1-1.

ot_jump: exact solution using JuMP, requires using JuMP, GLPK before it becomes available.

ot_convex: exact solution using Convex.jl, requires using Convex, GLPK before it becomes available.

Providing solver and options

options = (solver=sinkhorn_log!, tol=1e-6, iters=100_000)
distance = OptimalTransportRootDistance(domain=Continuous(), p=1, β=0.001)
SpectralDistances.evaluate(distance, model1, model2; options...)