Models and root manipulations

Models and root manipulations

Overview

This package supports two kind of LTI models, AR and ARMA. AR represents a model with only poles whereas ARMA has zeros as well.

To fit a model to data, one first has to specify a FitMethod, the options are

LS
PLR
TLS

For example, to estimate an AR model of order 30 using least-squares, we can do the following

julia> data = randn(1000);

julia> fitmethod = LS(na=30)
LS(30, 0.01)

julia> SpectralDistances.fitmodel(fitmethod, data)
AR{StaticArrays.SArray{Tuple{31},Float64,1,31},DiscreteRoots{Complex{Float64},Array{Complex{Float64},1}},ContinuousRoots{Complex{Float64},Array{Complex{Float64},1}}}([1.0, -0.02867099988927749, 0.0011823100005588668, 0.0172798668096625, 0.021040053322919545, 0.02544028345861928, -0.06192221361398596, -0.015448801252186766, 0.014026351110243843, 0.007113252793827176  …  -0.011196773463211076, 0.05426067814589986, 0.019102918773866707, -0.008150955353502017, -0.025018173472050105, 0.0010985235212486327, -0.01289301355775493, 0.008676211837759268, -0.004014609044867252, 0.02833878346298362], [1.0, 3.5635239721411733, 55.03055845237779, 169.8140097788977, 1319.9300073061002, 3538.303120822359, 18230.0842754238, 42494.208469140685, 161281.5403876998, 326533.5635190713  …  1.9541210794794742e7, 1.4931469419249842e7, 8.456861225264786e6, 4.7771146688061e6, 1.9830081136738465e6, 763126.8199467416, 206540.9982963334, 45369.29126210703, 6109.952769104339, 403.4925778932649], Complex{Float64}[-0.9142631615321836 - 0.10009904284811869im, -0.8406468369095961 - 0.29620588503332923im, -0.7616918829016566 - 0.4560702120141067im, -0.6488866496649718 - 0.6047275656444937im, -0.5028772949426371 - 0.7170221075510996im, -0.3326681228387137 - 0.7966334505312901im, -0.23487506356477045 - 0.8680631015971557im, 0.017053182679218598 - 0.9061639726559698im, 0.2540520821504141 - 0.8365566167390255im, 0.2803579048454399 - 0.7865570089903962im  …  0.2803579048454399 + 0.7865570089903962im, 0.2540520821504141 + 0.8365566167390255im, 0.017053182679218598 + 0.9061639726559698im, -0.23487506356477045 + 0.8680631015971557im, -0.3326681228387137 + 0.7966334505312901im, -0.5028772949426371 + 0.7170221075510996im, -0.6488866496649718 + 0.6047275656444937im, -0.7616918829016566 + 0.4560702120141067im, -0.8406468369095961 + 0.29620588503332923im, -0.9142631615321836 + 0.10009904284811869im], Complex{Float64}[-0.08367886947704442 - 3.0325409816258384im, -0.11506832888329814 - 2.802821625790217im, -0.11901784365312393 - 2.6020858225209342im, -0.1199228082754376 - 2.391405348866525im, -0.1326301472427963 - 2.1824219982143624im, -0.14698896863517444 - 1.96637602934195im, -0.10616372599080863 - 1.835042822681544im, -0.098357956291442 - 1.5519794565080807im, -0.13435202752331085 - 1.2759596252153371im, -0.180289503728299 - 1.2283986659251598im  …  -0.180289503728299 + 1.2283986659251598im, -0.13435202752331085 + 1.2759596252153371im, -0.098357956291442 + 1.5519794565080807im, -0.10616372599080863 + 1.835042822681544im, -0.14698896863517444 + 1.96637602934195im, -0.1326301472427963 + 2.1824219982143624im, -0.1199228082754376 + 2.391405348866525im, -0.11901784365312393 + 2.6020858225209342im, -0.11506832888329814 + 2.802821625790217im, -0.08367886947704442 + 3.0325409816258384im])

Type reference

SpectralDistances.AR

SpectralDistances.AR

SpectralDistances.ARMA

SpectralDistances.Continuous

SpectralDistances.ContinuousRoots

SpectralDistances.ContinuousRoots

SpectralDistances.Discrete

SpectralDistances.DiscreteRoots

SpectralDistances.DiscreteRoots

SpectralDistances.HungarianAssignement

SpectralDistances.SortAssignement

Function reference

SpectralDistances.AR

SpectralDistances.AR

SpectralDistances.ARMA

SpectralDistances.Continuous

SpectralDistances.ContinuousRoots

SpectralDistances.ContinuousRoots

SpectralDistances.Discrete

SpectralDistances.DiscreteRoots

SpectralDistances.DiscreteRoots

SpectralDistances.HungarianAssignement

SpectralDistances.SortAssignement

PolynomialRoots.roots

SpectralDistances.checkroots

SpectralDistances.coefficients

SpectralDistances.domain_transform

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.hungariansort

SpectralDistances.ls

SpectralDistances.plr

SpectralDistances.polar

SpectralDistances.reflect

SpectralDistances.residues

SpectralDistances.residues

SpectralDistances.residueweight

SpectralDistances.roots2poly

Docstrings

SpectralDistances.AR — Type.

struct AR{T} <: AbstractModel

Represents an all-pole transfer function, i.e., and AR model

Arguments:

a: denvec

ac: denvec cont. time

p: poles

source

SpectralDistances.AR — Type.

AR(X::AbstractArray, order::Int, λ=0.01)

Fit an AR model using LS as fitmethod

Arguments:

X: a signal

order: number of roots

λ: reg factor

source

SpectralDistances.ARMA — Type.

struct ARMA{T} <: AbstractModel

Represents an ARMA model, i.e., transfer function

Arguments:

c: numvec

cc: numvec cont. time

a: denvec

ac: denvec cont. time

z: zeros

p: poles

source

PolynomialRoots.roots — Function.

roots(m::AbstractModel)

Returns the roots of a model

source

SpectralDistances.checkroots — Method.

checkroots(r::DiscreteRoots) prints a warning if there are roots on the negative real axis.

source

SpectralDistances.coefficients — Method.

coefficients(::Domain, m::AbstractModel)

Return all fitted coefficients

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::LS, X::AbstractArray)

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::PLR, X::AbstractArray)

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::TLS, X::AbstractArray)

source

SpectralDistances.ls — Function.

ls(yA::AbstractTuple, λ=0.01)

Regularized Least-squares

source

SpectralDistances.plr — Method.

plr(y, na, nc, initial; λ=0.01)

Performs pseudo-linear regression to estimate an ARMA model.

Arguments:

y: signal

na: denomenator order

nc: numerator order

initial: fitmethod for the initial fit. Can be, e.g., LS, TLS or any function that returns a coefficient vector

λ: reg

source

SpectralDistances.residues — Function.

residues(a::AbstractVector, r=roots(reverse(a)))

Returns a vector of residues for the system represented by denominator polynomial a

source

SpectralDistances.residues — Method.

residues(r::AbstractRoots)

Returns a vector of residues for the system represented by roots r

source

SpectralDistances.roots2poly — Method.

roots2poly(roots)

Accepts a vector of complex roots and returns the polynomial with those roots

source

SpectralDistances.Continuous — Type.

Continuous time domain

source

SpectralDistances.ContinuousRoots — Type.

ContinuousRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represents roots in continuous time

source

SpectralDistances.ContinuousRoots — Method.

ContinuousRoots(r::DiscreteRoots) = begin

Represents roots of a polynomial in continuous time

source

SpectralDistances.Discrete — Type.

Discrete (sampled) time domain

source

SpectralDistances.DiscreteRoots — Type.

DiscreteRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represent roots in discrete time

source

SpectralDistances.DiscreteRoots — Method.

DiscreteRoots(r::ContinuousRoots) = begin

Represents roots of a polynomial in discrete time

source

SpectralDistances.HungarianAssignement — Type.

HungarianAssignement <: AbstractAssignmentMethod

Sort roots using Hungarian method

source

SpectralDistances.SortAssignement — Type.

SortAssignement{F} <: AbstractAssignmentMethod

Contains a single Function field that determines what to sort roots by.

source

SpectralDistances.domain_transform — Method.

domain_transform(d::Domain, e::AbstractRoots)

Change the domain of the roots

source

SpectralDistances.hungariansort — Function.

hungariansort(p1, p2)

takes two vectors of numbers and sorts and returns p2 such that it is in the order of the best Hungarian assignement between p1 and p2. Uses abs for comparisons, works on complex numbers.

source

SpectralDistances.polar — Method.

polar(e::Number)

magnitude and angle of a complex number

source

SpectralDistances.reflect — Method.

reflect(r::AbstractRoots)

Reflects unstable roots to a corresponding stable position (in unit circle for disc. in LHP for cont.)

source

SpectralDistances.residueweight — Method.

residueweight(e::AbstractRoots)

Returns a vector where each entry is roughly corresponding to the amount of energy contributed to the spectrum be each pole.

source