Models and root manipulations

Models and root manipulations

Overview

This package supports two kind of LTI models, AR and ARMA. AR represents a model with only poles whereas ARMA has zeros as well. These types are subtypes of ControlSystems.LTISystem, so many of the functions from the ControlSystems.jl toolbox work on these models as well. When acting like a ControlSystems.LTISystem, the default is to use the continuous-time representation of the model. The discrete-time representation can be obtained by tf(m, 1) where 1 is the sample time.

Note

This package makes the assumption that the sample time is 1 everywhere. When an AbstractModel is constructed, one must thus take care to rescale the frequency axis accordingly if this does not hold. If the discrete-time representation is never used, this is of no concern.

To fit a model to data, one first has to specify a FitMethod, the options are

LS
PLR
TLS

For example, to estimate an AR model of order 30 using least-squares, we can do the following

julia> data = randn(1000);

julia> fitmethod = LS(na=30)
LS(30, 0.01)

julia> SpectralDistances.fitmodel(fitmethod, data)
AR{StaticArrays.SArray{Tuple{31},Float64,1,31},DiscreteRoots{Complex{DoubleFloats.DoubleFloat{Float64}},Array{Complex{DoubleFloats.DoubleFloat{Float64}},1}},ContinuousRoots{Complex{DoubleFloats.DoubleFloat{Float64}},Array{Complex{DoubleFloats.DoubleFloat{Float64}},1}}}([1.0, -0.048760383468886975, -0.03589942374137859, -0.043658206838406224, 0.047769305535552475, 0.008294723131222, 0.00513152624991725, 0.012569965562439328, 0.009256452041902657, 0.04463033037202909  …  0.06294577486923049, -0.0013448607175151888, -0.01702076069262558, -0.0063863681550003016, 0.01215252849401347, 0.034006069407311074, -0.005380989229371497, 0.0589723267128616, 0.023661807763356885, 0.018884329612760258], [1.0, 3.9694228218454564, 56.69778262572079, 190.16356406003368, 1394.8889049702263, 3981.221195060724, 19697.021407448618, 48012.980024971366, 177728.34300585696, 370220.3564446817  …  2.1861047310042225e7, 1.810349496716726e7, 9.317546834401382e6, 5.85356791448221e6, 2.1398088186999457e6, 948624.4356270211, 218294.69472894177, 58878.98357831933, 6458.938272363693, 694.1895040637263], Complex{DoubleFloats.DoubleFloat{Float64}}[-0.8895645331456103 - 0.0996152615170458im, -0.8826769636953149 - 0.2929746334647734im, -0.7539432754355976 - 0.5157460336200533im, -0.6029551152397675 - 0.6473207349219342im, -0.48262495961224355 - 0.7315820300556449im, -0.24208229884782687 - 0.49644616742668685im, -0.2714108001045508 - 0.8680626577035404im, -0.06948328145721236 - 0.9045771092544794im, 0.17762730258112017 - 0.8804343401372411im, 0.3037462505822421 - 0.795644297082079im  …  0.3037462505822421 + 0.795644297082079im, 0.17762730258112017 + 0.8804343401372411im, -0.06948328145721236 + 0.9045771092544794im, -0.2714108001045508 + 0.8680626577035404im, -0.24208229884782687 + 0.49644616742668685im, -0.48262495961224355 + 0.7315820300556449im, -0.6029551152397675 + 0.6473207349219342im, -0.7539432754355976 + 0.5157460336200533im, -0.8826769636953149 + 0.2929746334647734im, -0.8895645331456103 + 0.0996152615170458im], Complex{DoubleFloats.DoubleFloat{Float64}}[-0.11079222182867658 - 3.0300751955236733im, -0.07254019423201034 - 2.821118214192737im, -0.0905065463416786 - 2.541641725948236im, -0.12258019849759282 - 2.32072470760581im, -0.13189220928287304 - 2.153960613481977im, -0.5936190788311814 - 2.0244994684829356im, -0.09485644267579743 - 1.8738293982894776im, -0.09734627771890887 - 1.6474587876926894im, -0.10739174342243896 - 1.371718997613788im, -0.1605769394667816 - 1.2061111036942922im  …  -0.1605769394667816 + 1.2061111036942922im, -0.10739174342243896 + 1.371718997613788im, -0.09734627771890887 + 1.6474587876926894im, -0.09485644267579743 + 1.8738293982894776im, -0.5936190788311814 + 2.0244994684829356im, -0.13189220928287304 + 2.153960613481977im, -0.12258019849759282 + 2.32072470760581im, -0.0905065463416786 + 2.541641725948236im, -0.07254019423201034 + 2.821118214192737im, -0.11079222182867658 + 3.0300751955236733im], 488.47965770674017)

Type reference

SpectralDistances.AR

SpectralDistances.AR

SpectralDistances.ARMA

SpectralDistances.Continuous

SpectralDistances.ContinuousRoots

SpectralDistances.ContinuousRoots

SpectralDistances.Discrete

SpectralDistances.DiscreteRoots

SpectralDistances.DiscreteRoots

SpectralDistances.HungarianAssignement

SpectralDistances.LS

SpectralDistances.PLR

SpectralDistances.SortAssignement

SpectralDistances.TLS

SpectralDistances.TimeDomain

Function reference

ControlSystems.denvec

ControlSystems.tf

ControlSystems.tf

PolynomialRoots.roots

SpectralDistances.checkroots

SpectralDistances.coefficients

SpectralDistances.domain_transform

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.hungariansort

SpectralDistances.ls

SpectralDistances.normalization_factor

SpectralDistances.normalize_energy

SpectralDistances.plr

SpectralDistances.polar

SpectralDistances.reflect

SpectralDistances.residues

SpectralDistances.residues

SpectralDistances.residueweight

SpectralDistances.roots2poly

SpectralDistances.spectralenergy

SpectralDistances.spectralenergy

Docstrings

SpectralDistances.AR — Type.

struct AR <: AbstractModel

Represents an all-pole transfer function, i.e., and AR model

Arguments:

a: denvec

ac: denvec cont. time

p: discrete time poles

pc: continuous time poles

b: Numerator scalar

source

SpectralDistances.AR — Method.

AR(X::AbstractArray, order::Int)

Fit an AR model using TLS as fitmethod

Arguments:

X: a signal

order: number of roots

source

SpectralDistances.ARMA — Type.

struct ARMA{T} <: AbstractModel

Represents an ARMA model, i.e., transfer function

Arguments:

c: numvec

cc: numvec cont. time

a: denvec

ac: denvec cont. time

z: zeros

p: poles

source

SpectralDistances.LS — Type.

LS <: FitMethod

Arguments:

na::Int: number of roots (order of the system)

λ::Float64 = 0.01: reg factor

source

SpectralDistances.PLR — Type.

PLR <: FitMethod

Pseudo linear regression. Estimates the noise components by performing an initial fit of higher order.

Arguments:

nc::Int: order of numerator

na::Int: order of denomenator

initial::T = TLS(na=80): fitmethod for the initial fit. Can be, e.g., LS, TLS or any function that returns a coefficient vector

λ::Float64 = 0.0001: reg factor

source

SpectralDistances.TLS — Type.

TLS <: FitMethod

Total least squares

Arguments:

na::Int: number of roots (order of the system)

source

PolynomialRoots.roots — Function.

roots(m::AbstractModel)

Returns the roots of a model

source

SpectralDistances.checkroots — Method.

checkroots(r::DiscreteRoots) prints a warning if there are roots on the negative real axis.

source

SpectralDistances.coefficients — Method.

coefficients(::Domain, m::AbstractModel)

Return all fitted coefficients

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::LS, X::AbstractArray)

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::PLR, X::AbstractArray)

source

SpectralDistances.fitmodel — Method.

fitmodel(fm::TLS, X::AbstractArray)

source

SpectralDistances.ls — Function.

ls(yA::AbstractTuple, λ=0.01)

Regularized Least-squares

source

SpectralDistances.normalization_factor — Method.

normalize_energy(r)

Returns the factor that, when used to multiply the poles, results in a system with unit spectral energy.

source

SpectralDistances.normalize_energy — Method.

normalize_energy(r)

Returns poles scaled to achieve unit spectral energy.

source

SpectralDistances.plr — Method.

plr(y, na, nc, initial; λ=0.01)

Performs pseudo-linear regression to estimate an ARMA model.

Arguments:

y: signal

na: denomenator order

nc: numerator order

initial: fitmethod for the initial fit. Can be, e.g., LS, TLS or any function that returns a coefficient vector

λ: reg

source

SpectralDistances.residues — Function.

residues(a::AbstractVector, r=roots(reverse(a)))

Returns a vector of residues for the system represented by denominator polynomial a

source

SpectralDistances.residues — Method.

residues(r::AbstractRoots)

Returns a vector of residues for the system represented by roots r

source

SpectralDistances.roots2poly — Method.

roots2poly(roots)

Accepts a vector of complex roots and returns the polynomial with those roots

source

SpectralDistances.spectralenergy — Method.

spectralenergy(G::LTISystem)

Calculates the energy in the spectrum associated with G

source

SpectralDistances.spectralenergy — Method.

spectralenergy(d::TimeDomain, a::AbstractVector, b)

Calculates the energy in the spectrum associated with denominator polynomial a

source

SpectralDistances.Continuous — Type.

Continuous time domain

source

SpectralDistances.ContinuousRoots — Type.

ContinuousRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represents roots in continuous time

source

SpectralDistances.ContinuousRoots — Method.

ContinuousRoots(r::DiscreteRoots) = begin

Represents roots of a polynomial in continuous time

source

SpectralDistances.Discrete — Type.

Discrete (sampled) time domain

source

SpectralDistances.DiscreteRoots — Type.

DiscreteRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represent roots in discrete time

source

SpectralDistances.DiscreteRoots — Method.

DiscreteRoots(r::ContinuousRoots) = begin

Represents roots of a polynomial in discrete time

source

SpectralDistances.HungarianAssignement — Type.

HungarianAssignement <: AbstractAssignmentMethod

Sort roots using Hungarian method

source

SpectralDistances.SortAssignement — Type.

SortAssignement{F} <: AbstractAssignmentMethod

Contains a single Function field that determines what to sort roots by.

source

SpectralDistances.TimeDomain — Type.

Abstract type that represents a time-domain, either Discrete or Continuous

source

SpectralDistances.domain_transform — Method.

domain_transform(d::Domain, e::AbstractRoots)

Change the domain of the roots

source

SpectralDistances.hungariansort — Function.

hungariansort(p1, p2)

takes two vectors of numbers and sorts and returns p2 such that it is in the order of the best Hungarian assignement between p1 and p2. Uses abs for comparisons, works on complex numbers.

source

SpectralDistances.polar — Method.

polar(e::Number)

magnitude and angle of a complex number

source

SpectralDistances.reflect — Method.

reflect(r::AbstractRoots)

Reflects unstable roots to a corresponding stable position (in unit circle for disc. in LHP for cont.)

source

SpectralDistances.residueweight — Method.

residueweight(e::AbstractRoots)

Returns a vector where each entry is roughly corresponding to the amount of energy contributed to the spectrum be each pole.

source

ControlSystems.tf — Method.

ControlSystems.tf(m::AR, ts)

Convert model to a transfer function compatible with ControlSystems.jl

source

ControlSystems.tf — Method.

ControlSystems.tf(m::AR)

Convert model to a transfer function compatible with ControlSystems.jl

source

ControlSystems.denvec — Method.

ControlSystems.denvec(::TimeDomain, m::AbstractModel)

Get the denominator polynomial vector

source