Models and root manipulations

Overview

This package supports two kind of LTI models, AR and ARMA. AR represents a model with only poles whereas ARMA has zeros as well. These types are subtypes of ControlSystems.LTISystem, so many of the functions from the ControlSystems.jl toolbox work on these models as well. When acting like a ControlSystems.LTISystem, the default is to use the continuous-time representation of the model. The discrete-time representation can be obtained by tf(m, 1) where 1 is the sample time.

Note

This package makes the assumption that the sample time is 1 everywhere. When an AbstractModel is constructed, one must thus take care to rescale the frequency axis accordingly if this does not hold. If the discrete-time representation is never used, this is of no concern.

To fit a model to data, one first has to specify a FitMethod, the options are

IRLS
LS
PLR
TLS

For example, to estimate an AR model of order 12 using least-squares, we can do the following

julia> data = randn(1000);

julia> fitmethod = LS(na=12)
LS(12, 0.01)

julia> SpectralDistances.fitmodel(fitmethod, data)
AR{StaticArrays.SArray{Tuple{13},Float64,1,13},DiscreteRoots{Complex{Double64},Array{Complex{Double64},1}},ContinuousRoots{Complex{Double64},Array{Complex{Double64},1}}}([1.0, -0.04693486779452964, -0.041241787322509445, -0.03983579203307957, 0.04090287048708235, 0.001441644676785205, 0.005136930970222201, 0.008859757552174163, 0.015811075487351268, 0.032960712025009865, 0.03684931207156168, -0.022463067975064414, 0.07256370327425937], [1.0, 2.6232904370705175, 22.349746327986406, 44.73555490478215, 172.50673789628567, 255.90473750024248, 561.2519551365834, 580.2724591968944, 745.8026466905417, 473.07952620424476, 323.2201687182947, 91.08608593380168, 20.72338551662842], Complex{Double64}[-0.7888891104608553 - 0.2252948980859896im, -0.5720135261390982 - 0.6017367584510012im, -0.18224621273111816 - 0.780633211409247im, 0.22902675416806015 - 0.7283629309001002im, 0.5402013864091196 - 0.5575407359005061im, 0.7973881426511567 - 0.2385733407796607im, 0.7973881426511567 + 0.2385733407796607im, 0.5402013864091196 + 0.5575407359005061im, 0.22902675416806015 + 0.7283629309001002im, -0.18224621273111816 + 0.780633211409247im, -0.5720135261390982 + 0.6017367584510012im, -0.7888891104608553 + 0.2252948980859896im], Complex{Double64}[-0.19792784859121945 - 2.8634125252393874im, -0.18604906424135984 - 2.3308765996576604im, -0.22111494337773194 - 1.800147876818909im, -0.2698133668717774 - 1.2661446412082842im, -0.25319340323171236 - 0.8011923169764876im, -0.18354659222145767 - 0.29071671232098323im, -0.18354659222145767 + 0.29071671232098323im, -0.25319340323171236 + 0.8011923169764876im, -0.2698133668717774 + 1.2661446412082842im, -0.22111494337773194 + 1.800147876818909im, -0.18604906424135984 + 2.3308765996576604im, -0.19792784859121945 + 2.8634125252393874im], 219.67895667195114)

Type reference

SpectralDistances.AR

SpectralDistances.AR

SpectralDistances.ARMA

SpectralDistances.AbstractModel

SpectralDistances.Continuous

SpectralDistances.ContinuousRoots

SpectralDistances.ContinuousRoots

SpectralDistances.Discrete

SpectralDistances.DiscreteRoots

SpectralDistances.DiscreteRoots

SpectralDistances.FitMethod

SpectralDistances.HungarianAssignement

SpectralDistances.IRLS

SpectralDistances.LS

SpectralDistances.PLR

SpectralDistances.SortAssignement

SpectralDistances.TLS

SpectralDistances.TimeDomain

Function reference

ControlSystems.denvec

ControlSystems.tf

ControlSystems.tf

PolynomialRoots.roots

SpectralDistances.checkroots

SpectralDistances.coefficients

SpectralDistances.domain_transform

SpectralDistances.examplemodels

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.fitmodel

SpectralDistances.hungariansort

SpectralDistances.ls

SpectralDistances.normalization_factor

SpectralDistances.normalize_energy

SpectralDistances.plr

SpectralDistances.polar

SpectralDistances.reflect

SpectralDistances.residues

SpectralDistances.residues

SpectralDistances.residueweight

SpectralDistances.roots2poly

SpectralDistances.spectralenergy

SpectralDistances.spectralenergy

Docstrings

SpectralDistances.AR — Type

struct AR <: AbstractModel

Represents an all-pole transfer function, i.e., and AR model

Arguments:

a: denvec

ac: denvec cont. time

p: discrete time poles

pc: continuous time poles

b: Numerator scalar

SpectralDistances.AR — Method

AR(X::AbstractArray, order::Int)

Fit an AR model using TLS as fitmethod

Arguments:

X: a signal

order: number of roots

SpectralDistances.ARMA — Type

struct ARMA{T} <: AbstractModel

Represents an ARMA model, i.e., transfer function

Arguments:

b: numvec

bc: numvec cont. time

a: denvec

ac: denvec cont. time

z: zeros

p: poles

SpectralDistances.AbstractModel — Type

abstract type AbstractModel <: ControlSystems.LTISystem end

SpectralDistances.FitMethod — Type

Abstract type that represents a way to fit a model to data

SpectralDistances.IRLS — Type

IRLS <: FitMethod

Iteratively reqeighted least squares. This fitmethod is currently not recommended, it does not appear to produce nice spectra. (feel free to try it out though).

Arguments:

na::Int: number of roots (order of the system)

SpectralDistances.LS — Type

LS <: FitMethod

This fitmethod is a good default option.

Arguments:

na::Int: number of roots (order of the system). The number of peaks in the spectrum will be na÷2.

λ::Float64 = 0.01: reg factor

SpectralDistances.PLR — Type

PLR <: FitMethod

Pseudo linear regression. Estimates the noise components by performing an initial fit of higher order. Tihs fitmethod produces an ARMA model. Support for ARMA models is not as strong as for AR models.

Arguments:

nc::Int: order of numerator

na::Int: order of denomenator

initial::T = TLS(na=80): fitmethod for the initial fit. Can be, e.g., LS, TLS or any function that returns a coefficient vector

λ::Float64 = 0.0001: reg factor

SpectralDistances.TLS — Type

TLS <: FitMethod

Total least squares. This fit method is good if the spectrum has sharp peaks, in particular if the number of peaks is known in advance.

Arguments:

na::Int: number of roots (order of the system). The number of peaks in the spectrum will be na÷2.

PolynomialRoots.roots — Function

roots(m::AbstractModel)

Returns the roots of a model

SpectralDistances.checkroots — Method

checkroots(r::DiscreteRoots) prints a warning if there are roots on the negative real axis.

SpectralDistances.coefficients — Method

coefficients(::Domain, m::AbstractModel)

Return all fitted coefficients

SpectralDistances.examplemodels — Function

examplemodels(n = 10)

Return n random models with 6 complex poles each.

SpectralDistances.fitmodel — Method

fitmodel(fm::IRLS, X::AbstractArray)

SpectralDistances.fitmodel — Method

fitmodel(fm::LS, X::AbstractArray)

SpectralDistances.fitmodel — Method

fitmodel(fm::PLR, X::AbstractArray)

SpectralDistances.fitmodel — Method

fitmodel(fm::TLS, X::AbstractArray)

SpectralDistances.ls — Function

ls(yA::AbstractTuple, λ=0.01)

Regularized Least-squares

SpectralDistances.normalization_factor — Method

normalize_energy(r)

Returns the factor that, when used to multiply the poles, results in a system with unit spectral energy.

SpectralDistances.normalize_energy — Method

normalize_energy(r)

Returns poles scaled to achieve unit spectral energy.

SpectralDistances.plr — Method

plr(y, na, nc, initial; λ=0.01)

Performs pseudo-linear regression to estimate an ARMA model.

Arguments:

y: signal

na: denomenator order

nc: numerator order

initial: fitmethod for the initial fit. Can be, e.g., LS, TLS or any function that returns a coefficient vector

λ: reg

SpectralDistances.residues — Function

residues(a::AbstractVector, b, r=roots(reverse(a)))

Returns a vector of residues for the system represented by denominator polynomial a

Ref: slide 21 https://stanford.edu/~boyd/ee102/rational.pdf Tihs methid is numerically sensitive. Note that if two poles are on top of each other, the residue is infinite.

SpectralDistances.residues — Method

residues(r::AbstractRoots)

Returns a vector of residues for the system represented by roots r

SpectralDistances.roots2poly — Method

roots2poly(roots)

Accepts a vector of complex roots and returns the polynomial with those roots

SpectralDistances.spectralenergy — Method

spectralenergy(G::LTISystem)

Calculates the energy in the spectrum associated with G

SpectralDistances.spectralenergy — Method

spectralenergy(d::TimeDomain, a::AbstractVector, b)

Calculates the energy in the spectrum associated with transfer function b/a

SpectralDistances.Continuous — Type

Continuous time domain

SpectralDistances.ContinuousRoots — Type

ContinuousRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represents roots in continuous time

SpectralDistances.ContinuousRoots — Method

ContinuousRoots(r::DiscreteRoots) = begin

Represents roots of a polynomial in continuous time

SpectralDistances.Discrete — Type

Discrete (sampled) time domain

SpectralDistances.DiscreteRoots — Type

DiscreteRoots{T, V <: AbstractVector{T}} <: AbstractRoots{T}

Represent roots in discrete time

SpectralDistances.DiscreteRoots — Method

DiscreteRoots(r::ContinuousRoots) = begin

Represents roots of a polynomial in discrete time

SpectralDistances.HungarianAssignement — Type

HungarianAssignement <: AbstractAssignmentMethod

Sort roots using Hungarian method

SpectralDistances.SortAssignement — Type

SortAssignement{F} <: AbstractAssignmentMethod

Contains a single Function field that determines what to sort roots by.

SpectralDistances.TimeDomain — Type

Abstract type that represents a time-domain, either Discrete or Continuous

SpectralDistances.domain_transform — Method

domain_transform(d::Domain, e::AbstractRoots)

Change the domain of the roots

SpectralDistances.hungariansort — Function

hungariansort(p1, p2)

takes two vectors of numbers and sorts and returns p2 such that it is in the order of the best Hungarian assignement between p1 and p2. Uses abs for comparisons, works on complex numbers.

SpectralDistances.polar — Method

polar(e::Number)

magnitude and angle of a complex number

SpectralDistances.reflect — Method

reflect(r::AbstractRoots)

Reflects unstable roots to a corresponding stable position (in unit circle for disc. in LHP for cont.)

SpectralDistances.residueweight — Method

residueweight(e::AbstractRoots)

Returns a vector where each entry is roughly corresponding to the amount of energy contributed to the spectrum be each pole.

ControlSystems.tf — Method

ControlSystems.tf(m::AR, ts)

Convert model to a transfer function compatible with ControlSystems.jl

ControlSystems.tf — Method

ControlSystems.tf(m::AR)

Convert model to a transfer function compatible with ControlSystems.jl

ControlSystems.denvec — Method

ControlSystems.denvec(::TimeDomain, m::AbstractModel)

Get the denominator polynomial vector